services/ms-cli-interface/requirements.txt
colorama==0.4.6
rich==13.6.0
python-dotenv==1.0.0
